Remove duplicated ir_serialization tests
