--- src/tests/functional/inference_engine/ir_serialization/models/add_abc_initializers.xml
@@ -1,93 +1,0 @@
-<?xml version="1.0" ?>
-<net name="add_abc_initializers" version="10">
-	<layers>
-		<layer id="0" name="add_node1/Output_0/Data__const" type="Const" version="opset1">
-			<data element_type="f32" offset="0" shape="2,2" size="16"/>
-			<output>
-				<port id="0" precision="FP32">
-					<dim>2</dim>
-					<dim>2</dim>
-				</port>
-			</output>
-		</layer>
-		<layer id="1" name="C" type="Parameter" version="opset1">
-			<data element_type="f32" shape="2,2"/>
-			<output>
-				<port id="0" precision="FP32">
-					<dim>2</dim>
-					<dim>2</dim>
-				</port>
-			</output>
-		</layer>
-		<layer id="2" name="Y" type="Add" version="opset1">
-			<input>
-				<port id="0">
-					<dim>2</dim>
-					<dim>2</dim>
-				</port>
-				<port id="1">
-					<dim>2</dim>
-					<dim>2</dim>
-				</port>
-			</input>
-			<output>
-				<port id="2" precision="FP32">
-					<dim>2</dim>
-					<dim>2</dim>
-				</port>
-			</output>
-		</layer>
-		<layer id="3" name="Y/sink_port_0" type="Result" version="opset1">
-			<input>
-				<port id="0">
-					<dim>2</dim>
-					<dim>2</dim>
-				</port>
-			</input>
-		</layer>
-	</layers>
-	<edges>
-		<edge from-layer="0" from-port="0" to-layer="2" to-port="0"/>
-		<edge from-layer="1" from-port="0" to-layer="2" to-port="1"/>
-		<edge from-layer="2" from-port="2" to-layer="3" to-port="0"/>
-	</edges>
-	<meta_data>
-		<MO_version value="custom_master_29f1c38ba0ae51897f47946d79d6bd6be7a494f0"/>
-		<cli_parameters>
-			<caffe_parser_path value="DIR"/>
-			<data_type value="float"/>
-			<disable_nhwc_to_nchw value="False"/>
-			<disable_omitting_optional value="False"/>
-			<disable_resnet_optimization value="False"/>
-			<disable_weights_compression value="False"/>
-			<enable_concat_optimization value="False"/>
-			<enable_flattening_nested_params value="False"/>
-			<enable_ssd_gluoncv value="False"/>
-			<extensions value="DIR"/>
-			<framework value="onnx"/>
-			<freeze_placeholder_with_value value="{}"/>
-			<generate_deprecated_IR_V7 value="False"/>
-			<input_model value="DIR/add_abc_initializers.onnx"/>
-			<input_model_is_text value="False"/>
-			<k value="DIR/CustomLayersMapping.xml"/>
-			<keep_shape_ops value="True"/>
-			<legacy_mxnet_model value="False"/>
-			<log_level value="ERROR"/>
-			<mean_scale_values value="{}"/>
-			<mean_values value="()"/>
-			<model_name value="add_abc_initializers"/>
-			<output_dir value="DIR"/>
-			<placeholder_data_types value="{}"/>
-			<progress value="False"/>
-			<remove_memory value="False"/>
-			<remove_output_softmax value="False"/>
-			<reverse_input_channels value="False"/>
-			<save_params_from_nd value="False"/>
-			<scale_values value="()"/>
-			<silent value="False"/>
-			<static_shape value="False"/>
-			<stream_output value="False"/>
-			<unset unset_cli_parameters="batch, counts, disable_fusing, disable_gfusing, finegrain_fusing, input, input_checkpoint, input_meta_graph, input_proto, input_shape, input_symbol, mean_file, mean_file_offsets, move_to_preprocess, nd_prefix_name, output, placeholder_shapes, pretrained_model_name, saved_model_dir, saved_model_tags, scale, tensorboard_logdir, tensorflow_custom_layer_libraries, tensorflow_custom_operations_config_update, tensorflow_object_detection_api_pipeline_config, tensorflow_use_custom_operations_config, transformations_config"/>
-		</cli_parameters>
-	</meta_data>
-</net>
